Refactor: #24 management 관련 코드 패키지 변경
