+++ src/main/java/com/sparta/parknav/management/controller/MgtController.java
@@ -1,13 +1,13 @@
-package com.sparta.parknav.ticket.controller;
+package com.sparta.parknav.management.controller;
 
 import com.sparta.parknav.global.response.ApiResponseDto;
 import com.sparta.parknav.global.security.AdminDetailsImpl;
 import com.sparta.parknav.global.security.UserDetailsImpl;
-import com.sparta.parknav.ticket.dto.request.CarNumRequestDto;
-import com.sparta.parknav.ticket.dto.response.CarInResponseDto;
-import com.sparta.parknav.ticket.dto.response.CarOutResponseDto;
-import com.sparta.parknav.ticket.dto.response.ParkMgtResponseDto;
-import com.sparta.parknav.ticket.service.MgtService;
+import com.sparta.parknav.management.dto.request.CarNumRequestDto;
+import com.sparta.parknav.management.dto.response.CarInResponseDto;
+import com.sparta.parknav.management.dto.response.CarOutResponseDto;
+import com.sparta.parknav.management.dto.response.ParkMgtResponseDto;
+import com.sparta.parknav.management.service.MgtService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
