+++ src/main/java/com/sparta/parknav/management/dto/response/ParkMgtResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.parknav.ticket.dto.response;
+package com.sparta.parknav.management.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
